implemented backend for couple of views

+++ PlagiarismChecker/PlagiarismChecker/Checker.cs
@@ -8,11 +8,11 @@ namespace PlagiarismChecker.App_Code
     public class Checker
     {
         public static List<Dictionary<string, int>> Documents { get; set; } = new List<Dictionary<string, int>>();
-        public static double MesaureSimilarity(Dictionary<string, int> doc1, Dictionary<string, int> doc2)
+        public static float MesaureSimilarity(Dictionary<string, int> doc1, Dictionary<string, int> doc2)
         {
-            double similarity = 0;
-            double magDoc1 = getDocumentMagnitude(doc1);
-            double magDoc2 = getDocumentMagnitude(doc2);
+            float similarity = 0;
+            float magDoc1 = getDocumentMagnitude(doc1);
+            float magDoc2 = getDocumentMagnitude(doc2);
             foreach (KeyValuePair<string, int> valueDoc1 in doc1)
             {
                 if (doc2.ContainsKey(valueDoc1.Key) == true)
@@ -21,18 +21,18 @@ namespace PlagiarismChecker.App_Code
 
             return similarity / (magDoc2 * magDoc1);
         }
-        private static double getDocumentMagnitude(Dictionary<string, int> doc)
+        private static float getDocumentMagnitude(Dictionary<string, int> doc)
         {
-            double magnitude = 0;
+            float magnitude = 0;
             foreach (KeyValuePair<string, int> valueDoc in doc)
             {
                 magnitude += calcTermIDF(valueDoc.Key)*(valueDoc.Value * valueDoc.Value);
             }
-            return Math.Sqrt(magnitude);
+            return (float)Math.Sqrt((double)magnitude);
         }
-        private static double calcTermIDF(string term)
+        private static float calcTermIDF(string term)
         {
-            double foundinDocs = 0;
+            int foundinDocs = 0;
 
             if (Documents.Count == 0)
                 return 1;
@@ -45,7 +45,7 @@ namespace PlagiarismChecker.App_Code
             if (foundinDocs == 0)
                 return 1;
 
-            double idf = Math.Log10((double)Documents.Count / foundinDocs);
+            float idf = (float)Math.Log10((double)Documents.Count / foundinDocs);
             if (idf == 0)
             {
                 return 0.01f;
